refactor(payments): reorganize component imports for better structure
